requirements.txt
scipy>=1.4.1
torchvision>=0.6.1
opencv_python>=4.3.0.36
coloredlogs>=14.0
pandas>=1.0.3
numpy>=1.17.4
GPSPhoto>=2.2.3
Pillow>=7.2.0
click>=7.1.2
piexif>=1.1.3
exifread
requests
